+++ docs/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add upstream references to Jira trackers on creation (OSIDB-3148)
 - Change ACL mixin serializer to support internal ACLs (OSIDB-3578)
 - Make product definitions collector atomic (OSIDB-3590)
+- Validate that the CVSSv3 score is zero for flaws with impact "None" (OSIDB-3581)
 
 ## [4.5.2] - 2024-10-24
 ### Changed